various refactoring such as putting persistence related functions into sub dir
