+++ shared/src/persistence/dynamo/cases/saveCase.test.js
@@ -10,6 +10,7 @@ const applicationContext = {
   environment: {
     stage: 'local',
   },
+  getDynamoClient: () => client,
   filterCaseMetadata: ({ cases }) => cases,
   isAuthorizedForWorkItems: () => true,
 };
