+++ shared/src/persistence/dynamo/users/getUserById.js
