+++ shared/src/persistence/s3/getDocument.test.js
@@ -1,0 +1,32 @@
+const { getDocument } = require('./getDocument');
+
+describe('getDocument', () => {
+  it('returns the expected documentId after when the upload was successful', async () => {
+    const BLOB_DATA = 'abc';
+    let applicationContext = {
+      getBaseUrl: () => 'http://localhost',
+      getHttpClient: () => {
+        const fun = () => ({
+          data: BLOB_DATA,
+        });
+        fun.get = () => ({
+          data: 'http://localhost',
+        });
+        return fun;
+      },
+      getPersistenceGateway: () => ({
+        uploadPdf: () => BLOB_DATA,
+      }),
+      getCurrentUser: () => {
+        return { userId: 'taxpayer', role: 'petitioner' };
+      },
+      getCurrentUserToken: () => {
+        return '';
+      },
+    };
+    const result = await getDocument({
+      applicationContext,
+    });
+    expect(result).toEqual(new Blob([BLOB_DATA], { type: 'application/pdf' }));
+  });
+});
