+++ shared/src/persistence/s3/getUploadPolicy.js
