+++ web-client/src/presenter/actions/createTokenAction.js
@@ -1,7 +1,7 @@
 import jwt from 'jsonwebtoken';
 import { state } from 'cerebral';
 
-import { userMap } from '../../../../shared/src/persistence/getUserById';
+import { userMap } from '../../../../shared/src/persistence/dynamo/users/getUserById';
 
 export default async ({ get, props }) => {
   const name = props.token || get(state.form.name);
